# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Calculator Service","description":"Service for multiplying numbers, a Goa teaser","version":"1.0"},"servers":[{"url":"http://localhost:8000"}],"paths":{"/add/{a}/{b}":{"get":{"tags":["calc"],"summary":"add calc","operationId":"calc#add","parameters":[{"name":"a","in":"path","description":"Left operand","required":true,"schema":{"type":"integer","description":"Left operand","example":2918713324217948082,"format":"int64"},"example":6487313515641330478},{"name":"b","in":"path","description":"Right operand","required":true,"schema":{"type":"integer","description":"Right operand","example":7083893185661257248,"format":"int64"},"example":6467796626875016509}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":8410464532698730620,"format":"int64"},"example":7509468512202955456}}}}}},"/divide/{a}/{b}":{"get":{"tags":["calc"],"summary":"divide calc","operationId":"calc#divide","parameters":[{"name":"a","in":"path","description":"Left operand","required":true,"schema":{"type":"integer","description":"Left operand","example":9093376179136310938,"format":"int64"},"example":8851621330472964370},{"name":"b","in":"path","description":"Right operand","required":true,"schema":{"type":"integer","description":"Right operand","example":8131166797256069893,"format":"int64"},"example":2303530046912946382}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"number","example":0.5022122856443563,"format":"double"},"example":0.7436789905469807}}}}}},"/multiply/{a}/{b}":{"get":{"tags":["calc"],"summary":"multiply calc","operationId":"calc#multiply","parameters":[{"name":"a","in":"path","description":"Left operand","required":true,"schema":{"type":"integer","description":"Left operand","example":4416510739546570040,"format":"int64"},"example":5088516426134668496},{"name":"b","in":"path","description":"Right operand","required":true,"schema":{"type":"integer","description":"Right operand","example":2817649694697734565,"format":"int64"},"example":8220311738705198664}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":4353119450067006442,"format":"int64"},"example":7969322314291156173}}}}}},"/notes/{userID}":{"get":{"tags":["calc"],"summary":"getNotes calc","operationId":"calc#getNotes","parameters":[{"name":"userID","in":"path","description":"The email of the user","required":true,"schema":{"type":"string","description":"The email of the user","example":"Fugit a nulla."},"example":"Sed ut beatae sint illo quam."}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetNotesResponseBody"},"example":{"notes":[{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."},{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."}]}}}}}},"post":{"tags":["calc"],"summary":"createNote calc","operationId":"calc#createNote","parameters":[{"name":"userID","in":"path","description":"The UserID for the note","required":true,"schema":{"type":"string","description":"The UserID for the note","example":"Rerum est nostrum qui aut vero."},"example":"Labore dolores eveniet."}],"requestBody":{"description":"The Note data to be saved","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Note"},"example":{"Body":"Accusantium culpa odit eaque.","Title":"Veritatis id iure."}}}},"responses":{"201":{"description":"Created response."},"400":{"description":"Bad Request response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"},"example":{"id":"3F1FKVRR","message":"Value of ID must be an integer","name":"bad_request"}}}}}}},"/openapi.json":{"get":{"tags":["calc"],"summary":"Download ./gen/http/openapi.json","operationId":"calc#/openapi.json","responses":{"200":{"description":"File downloaded"}}}},"/subtract/{a}/{b}":{"get":{"tags":["calc"],"summary":"subtract calc","operationId":"calc#subtract","parameters":[{"name":"a","in":"path","description":"Left operand","required":true,"schema":{"type":"integer","description":"Left operand","example":7526834330431756466,"format":"int64"},"example":4000321415748316517},{"name":"b","in":"path","description":"Right operand","required":true,"schema":{"type":"integer","description":"Right operand","example":2820419433035539482,"format":"int64"},"example":3028519962502007746}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":5933435596807726301,"format":"int64"},"example":7770762273010691493}}}}}}},"components":{"schemas":{"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"example":{"id":"3F1FKVRR","message":"Value of ID must be an integer","name":"bad_request"},"required":["name","id","message","temporary","timeout","fault"]},"GetNotesResponseBody":{"type":"object","properties":{"notes":{"type":"array","items":{"$ref":"#/components/schemas/Note"},"description":"list of notes","example":[{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."},{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."},{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."},{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."}]}},"example":{"notes":[{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."},{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."},{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."},{"Body":"Quibusdam eius.","Title":"Necessitatibus quis quasi est eum ad."}]}},"Note":{"type":"object","properties":{"Body":{"type":"string","description":"The Body of the Note","example":"Omnis debitis voluptatum est enim possimus exercitationem."},"Title":{"type":"string","description":"The title of the Note","example":"Illum ea velit."}},"example":{"Body":"Excepturi et molestiae aperiam corporis dolore a.","Title":"Velit consequuntur eum illum voluptatem possimus."}}}},"tags":[{"name":"calc","description":"The calc service performs operations on numbers."}]}